+++ src/components/input/Input.stories.ts
@@ -2,6 +2,7 @@ import { boolean, text, select } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
 import Vue, { PropType } from 'vue';
 import WvuiInput from './Input.vue';
+import WvuiButton from '../button/Button.vue';
 import { InputType } from './InputType';
 import './Input.stories.less';
 
@@ -41,7 +42,7 @@ export const configurable = (): Vue.Component =>
 				:disabled="disabled"
 				@input="input"
 				@change="change"
-				@focus="focus" 
+				@focus="focus"
 				@blur="blur"
 			/>
 		</div>
@@ -68,6 +69,49 @@ export const withIcon = (): Vue.Component =>
 				icon="search"
 				:disabled="disabled"
 			/>
+		</div>
+	`
+	} );
+
+export const withButton = (): Vue.Component =>
+	Vue.extend( {
+		components: { WvuiInput, WvuiButton },
+		data() {
+			return {
+				InputType
+			};
+		},
+		template: `
+		<div class="sb-input-preview">
+			<wvui-input
+					placeholder="Search..."
+					:type="InputType.Search"
+				
+			>
+				<wvui-button action="progressive">Search</wvui-button>
+			</wvui-input>
+		</div>
+	`
+	} );
+
+export const wikipediaSearchInput = (): Vue.Component =>
+	Vue.extend( {
+		components: { WvuiInput, WvuiButton },
+		data() {
+			return {
+				InputType
+			};
+		},
+		template: `
+		<div class="sb-input-preview">
+			<wvui-input
+				placeholder="Search..."
+				:type="InputType.Search"
+        icon="search"
+					
+			>
+				<wvui-button>Search</wvui-button>
+			</wvui-input>
 		</div>
 	`
 	} );